Consolidate redundant code in train.py and train_curriculum.py into src/utils.py
